[refactor] Extract top harness generation to separate module. [python/assassyn/codegen/verilog/top.py] Create new module with generate_top_harness function extracted from CIRCTDumper._generate_top_harness. [python/assassyn/codegen/verilog/design.py] Remove _generate_top_harness method and import generate_top_harness from top module. Remove unused imports. This refactoring improves code organization by separating the complex top-level harness generation logic into its own dedicated module.
